+++ BTCPayServer.Plugins.LNbank/Hooks/AuthorizationRequirementHandler.cs
@@ -41,6 +41,7 @@ public class AuthorizationRequirementHandler : IPluginHookFilter
             {
                 UserId = new[] { userId },
                 WalletId = new[] { walletId },
+                IncludeTransactions = true,
                 IsServerAdmin = isServerAdmin
             });
         }