+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Admin overview: See how many wallets there are and what balances they have. (#16)
+- Admin can delete transactions.
 
 ### Changed